move BPartner/Product classes de.metas.bpartner_product
https://github.com/metasfresh/metasfresh/issues/5098
