+++ de.metas.swat/de.metas.swat.base/src/main/java/de/metas/order/process/impl/CreatePOFromSOsAggregationKeyBuilder.java
@@ -11,8 +11,8 @@ import org.compiere.model.I_C_Order;
 import org.compiere.model.I_C_OrderLine;
 import org.compiere.model.I_M_Product;
 
+import de.metas.bpartner_product.IBPartnerProductDAO;
 import de.metas.i18n.IMsgBL;
-import de.metas.purchasing.api.IBPartnerProductDAO;
 import de.metas.util.ILoggable;
 import de.metas.util.Loggables;
 import de.metas.util.Services;
