+++ de.metas.vertical.pharma.msv3.server-peer-metasfresh/src/main/java/de/metas/vertical/pharma/msv3/server/peer/metasfresh/services/MSV3StockAvailabilityService.java
@@ -19,6 +19,7 @@ import com.google.common.collect.ImmutableList;
 
 import ch.qos.logback.classic.Level;
 import de.metas.bpartner.BPartnerId;
+import de.metas.bpartner_product.IBPartnerProductDAO;
 import de.metas.document.sequence.IDocumentNoBuilderFactory;
 import de.metas.logging.LogManager;
 import de.metas.material.cockpit.stock.StockDataAggregateItem;
@@ -29,7 +30,6 @@ import de.metas.material.event.stock.StockChangedEvent;
 import de.metas.product.IProductDAO;
 import de.metas.product.ProductCategoryId;
 import de.metas.product.ProductId;
-import de.metas.purchasing.api.IBPartnerProductDAO;
 import de.metas.util.Check;
 import de.metas.util.GuavaCollectors;
 import de.metas.util.Loggables;
